app/requirements.txt
fastapi
uvicorn
pymongo
boto3
python-dotenv
passlib[bcrypt]
